backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
openai
google-generativeai
python-dotenv
pytest
pytest-asyncio
